backend/requirements.txt
Flask==3.0.3
waitress==2.1.2
requests==2.31.0
finnhub-python==2.4.21
google-genai
google-generativeai
flask_cors==5.0.1
pyjwt==2.10.1
gunicorn
